% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_gls_residuals.R
\name{analyse_gls_residuals}
\alias{analyse_gls_residuals}
\title{Residual analysis for GLS models}
\usage{
analyse_gls_residuals(gls_object)
}
\arguments{
\item{gls_object}{A 'gls' model object generated by the 'nlme'
package.}
}
\description{
This function evaluates residuals from 'gls' model objects.
It returns #' a normality plot, residuals vs. fitted values,
and a lag plot with a trend line
(residuals vs. lagged residuals) to evaluate autocorrelation. It also
provides a Shapiro-Wilk normality test, Levene's test for
homogeneity of variance, the Durbin-Watson test
for autocorrelation, and a t-test for zero mean.
}
\examples{
# Load required libraries
library("car")
library("nlme")

# Fit a GLS model
data(Orthodont)
model <- gls(distance ~ age, data = Orthodont, correlation = corAR1(form = ~ 1 | Subject))

# Analyze GLS residuals
analyse_gls_residuals(model)

}
