% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_lm_residuals.R
\name{analyse_lm_residuals}
\alias{analyse_lm_residuals}
\alias{analyze_lm_residuals}
\title{Residual analysis for linear models}
\usage{
iris_lm <- lm(iris$Sepal.Length, iris$Sepal.Width)
analyse_lm_residuals(iris_lm)
}
\arguments{
\item{lm_object}{An object of class 'lm'}
}
\value{
Residual plots: normality, residuals vs. fitted data,
autocorrelation (lagged residuals vs. residuals with fitted linear model).

Prints residual tests: Mean zero, Shapiro-Wilk normality test,
Levene Test for homogeneity of variance, and Run's test for randomness
}
\description{
Residual analysis for linear models
}
