% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_variance.R
\name{tests_variance}
\alias{tests_variance}
\title{Performs Levene tests and makes boxplots and density plot}
\usage{
tests_variance(Sepal.Length, Species)
}
\arguments{
\item{df}{A dataframe}

\item{data_var}{Variable of interest}

\item{group_var}{Grouping variable}
}
\value{
A plot grid with boxplot, boxplot with dots, and density plot.
Prints Levene test results.
}
\description{
Performs Levene tests and makes boxplots and density plot
}
