% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_norm.R
\name{tests_norm}
\alias{tests_norm}
\title{Performs Shapiro-Wilk test and makes normality plot (qq), density plot,
histogram, and histogram with normality curve.}
\usage{
tests_norm(data, variable)
}
\arguments{
\item{data}{A dataframe}

\item{variable}{Variable name passed as a string}
}
\value{
A plot grid with normality plot, histogram, density plot, and
histogram with normality curve. Prints Shapiro-Wilke test results.
}
\description{
Performs Shapiro-Wilk test and makes normality plot (qq), density plot,
histogram, and histogram with normality curve.
}
